+++ .circleci/integration_test.sh
@@ -41,4 +41,16 @@ fi
 
 grep --silent "== Warnings: 0 (not including reflection warnings)  Exceptions thrown: 0" output  || exit 1
 
+git submodule update --init --recursive
+
+cd ../core.async || exit 1
+
+if lein with-profile +test update-in :plugins conj "[jonase/eastwood \"RELEASE\"]" -- eastwood > output; then
+  echo "Should have failed! Emitted output:"
+  cat output
+  exit 1
+fi
+
+grep --silent "== Warnings: 39 (not including reflection warnings)  Exceptions thrown: 0" output  || exit 1
+
 exit 0