+++ test-third-party-deps/testcases/core_async_example.clj
@@ -1,0 +1,16 @@
+(ns testcases.core-async-example
+  (:require
+   [clojure.core.async :refer [go]]))
+
+(defprotocol P
+  (x [p]))
+
+;; taken from the core.async test suite
+(defrecord R [z]
+  P
+  (x [this]
+    (go
+      (loop []
+        (if (zero? (rand-int 3))
+          [z (.z this)]
+          (recur))))))