+++ changes.md
@@ -2,6 +2,13 @@
 
 ## Changes from 0.5.2 to
 
+#### New
+
+* Omit exceptions coming from individual top-level forms, if they made use of the `clojure.core.async/go` macro.
+  * Closes https://github.com/jonase/eastwood/issues/395
+  * This makes Eastwood more capable of analyzing core.async -based projects. It's a temporary measure though, as it ideally this would be simply fixed in the tools.analyzer project. 
+  * You can revert to the old behavior (which most likely will result in errors by passing `:abort-on-core-async-exceptions? true`) as a top-level Eastwood option. 
+
 #### Bugfixes
 
 * Fix the `eastwood.lint` `-main` program when invoked with no arguments.